+++ src/actions/client/goto-transactions-scroller-by-statement-request.ts
@@ -1,18 +1,23 @@
 import { checkOutdatedStatement } from 'actions/client/check-outdated-statement';
-import { ACTION } from 'interfaces/client';
+import type { ACTION } from 'interfaces/client';
+import type { ICreateRequestStatementDto } from 'interfaces/dto';
+import type { ENTRY_SOURCE_VIEW } from 'stream-constants';
+import { COMMON_STREAM_URL } from 'stream-constants/client';
 import { fatalHandler } from 'utils';
 import { singleAction, to } from '@platform/core';
 import type { IActionConfig } from '@platform/services';
-import type { IBaseEntity } from '@platform/services/client';
 import type { context } from './executor';
-import { gotoTransactionsScroller } from './goto-transactions-scroller';
 
 /** Действие перехода на скроллер проводок, по сущности "Запрос выписки". */
 export const gotoTransactionsScrollerByStatementRequest: IActionConfig<typeof context, Promise<void>> = {
-  action: ({ done, fatal, execute }, { service, showLoader, hideLoader }) => async ([doc]: [IBaseEntity]) => {
+  action: ({ done, fatal, execute }, { service, showLoader, hideLoader, router }) => async (
+    [doc]: [ICreateRequestStatementDto],
+    action: ACTION,
+    entrySourceView?: typeof ENTRY_SOURCE_VIEW
+  ) => {
     const {
       succeeded: [isOutdated],
-    } = await execute(checkOutdatedStatement, [doc], ACTION.VIEW);
+    } = await execute(checkOutdatedStatement, [doc], action);
 
     if (isOutdated) {
       done();
@@ -29,7 +34,9 @@ export const gotoTransactionsScrollerByStatementRequest: IActionConfig<typeof co
     fatal(res?.error);
     fatal(err);
 
-    await execute(gotoTransactionsScroller, [res?.data]);
+    if (res) {
+      router.push(`${COMMON_STREAM_URL.STATEMENT_TRANSACTIONS}/${res.data.id}`, { entrySourceView });
+    }
 
     done();
   },